Return indicies of type int

+++ generate_stimuli.py
@@ -455,7 +455,7 @@ def get_background_tiles_locations(frag_len, img_len, row_offset, space_bw_tiles
     start_y = np.array(start_y)
     start_y = np.reshape(start_y, (start_y.shape[0] * start_y.shape[1]))
 
-    loc_arr = np.array([start_x, start_y])
+    loc_arr = np.array([start_x, start_y]).astype(int)
     loc_arr = loc_arr.T
 
     return loc_arr
@@ -842,9 +842,9 @@ if __name__ == "__main__":
         tgt_n_visual_rf_start=image_size[0] // 2 - (full_tile_size[0] // 2)
     )
 
-    test_image = highlight_tiles(test_image, fragment_size, bg_tiles.astype(int), edge_color=(255, 255, 0))
-    test_image = highlight_tiles(test_image, full_tile_size, full_tile_starts.astype(int), edge_color=(255, 0, 0))
-    test_image = highlight_tiles(test_image, fragment_size, path_fragment_starts.astype(int), edge_color=(0, 0, 255))
+    test_image = highlight_tiles(test_image, fragment_size, bg_tiles, edge_color=(255, 255, 0))
+    test_image = highlight_tiles(test_image, full_tile_size, full_tile_starts, edge_color=(255, 0, 0))
+    test_image = highlight_tiles(test_image, fragment_size, path_fragment_starts, edge_color=(0, 0, 255))
 
     plt.figure()
     plt.imshow(test_image)
